aws-security-scanner/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
click>=8.1.0
jinja2>=3.1.0
pyyaml>=6.0
tabulate>=0.9.0
colorama>=0.4.6
networkx>=3.0
flask>=3.0.0
flask-socketio>=5.3.0
